Go back to importing "./js/three.js" directly
Apparently Webview doesn't play well with the <script type="module"> attribute. I made it work just fine in the browser, but it won't work on webkitgtk. :(

I'll keep it this way for now, until I find a way to import Three.js directly into Deno, and bundle it for webkit :D