Fixes #3 Bundling Works!!!! Finally!
I managed to make the bundling work as I wanted. So now all we need to do is to bundle with ./build.js

I also cleaned up index.html a bit. <script> no longer needs to have the type=module attribute
Also, we no longer need to import three.js inside the html. Everything is in the bundled app.js